refactor(Session): session 패키지 생성
