+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -1,8 +1,8 @@
 package nextstep.org.apache.coyote.http11;
 
-import org.apache.catalina.Manager;
-import org.apache.catalina.Session;
-import org.apache.catalina.SessionManager;
+import org.apache.session.Manager;
+import org.apache.session.Session;
+import org.apache.session.SessionManager;
 import org.apache.coyote.http11.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
